requirements.txt
torch>=1.8.0
pyro-ppl>=1.8.0
numpy>=1.19.0
matplotlib>=3.3.0
pandas>=1.1.0
transformers>=4.5.0
datasets>=1.5.0
scikit-learn>=0.24.0
tqdm>=4.45.0
seaborn>=0.11.0
ipython>=7.0.0
jupyter>=1.0.0 